app/backend/pyproject.toml
[project]
name = "financial-analysis-app"
version = "0.2.0"
description = "Django REST API for analyzing financial data from CSV and PDF documents"
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14.0,<4.0",
    "django-cors-headers>=4.3.1,<5.0",
    "django-filter>=24.0,<25.0",
    "pandas>=2.2.0,<3.0",
    "python-dateutil>=2.8.2,<3.0",
    "python-dotenv>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.0.0,<26.0",
    "flake8>=7.0.0,<8.0",
    "isort>=5.13.0,<6.0",
    "ipython>=8.20.0,<9.0",
    "pytest>=7.0.0,<8.0",
    "pytest-django>=4.7.0,<5.0",
    "pytest-cov>=4.0.0,<5.0",
    "factory-boy>=3.3.0,<4.0",
]

[tool.black]
line-length = 100
target-version = ['py314']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = ["migrations", ".venv"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "financial_analysis.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]
addopts = "--strict-markers --verbose"

[tool.uv.workspace]
members = [
    "app/backend",
]
